hackTonNettside/self_healing_swarm_ai/requirements.txt

# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Optional: Deep Learning (if using neural networks)
# torch>=2.0.0
# tensorflow>=2.13.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
